# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Sutas Personel",
   

    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "sutaspersonel.herokuapp.com",
...
  "paths": {
...
    "/news/search/{newsId}": {
      "get": {
        "tags": ["News"],
        "summary": "News search",
        "description": "",

        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/News"
            }
          }
        },
        "parameters": [
          { "required": true, "type": "string", "name": "newsId", "in": "path" }
        ]
      }
    },
...
    "/news/update/{newsId}": {
      "patch": {
        "tags": ["News"],
        "summary": "News update",
        "description": "",

        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/News"
            }
          }
        },
        "parameters": [
          { "required": true, "type": "string", "name": "newsId", "in": "path" }
        ]
      }
    },
    "/news/add/": {
      "post": {
        "tags": ["News"],
        "summary": "News add",
        "description": "",

        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/News"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": { "$ref": "#/definitions/News" }
          }
        ]
      }
    },
...
  },
  "definitions": {
    "News": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "date": {
          "type": "string",
          "format": "date"
        }
      }
    },
...
  }
}